Support MetaStore without config file
